TG_MINI_APP/requirements.txt
fastapi
uvicorn
tortoise-orm
asyncpg
aerich
python-dotenv
aiogram
python-multipart